seperate files for seperate reactors
